changed config setups, removed protein destructor
changed configurating setters for all the classes. Changed the docstirngs respectively. Recoded the 'create pathways method of the Protein class with regard to indexing and method parameters. Removed the class destructor because the garbage collector was causing an error with regard to resource clean up done by the CM. I don't know why that was happening as I was setting the cleaned_up flag to true every time.'